Added --lua-dir, --bitmap-dir, and --config-file command line options. Added native 64 bit support. Added option to load high-quality MP3 background music rather than MIDI. Improved error reporting when asynchronously loading music. Fixed crash resulting from music data being garbage collected while still in use.

+++ CorsixTH/Src/rnc.cpp
@@ -316,7 +316,7 @@ static int rnc_unpack(const unsigned char* input, unsigned char* output)
      * Check the packed-data CRC. Also save the unpacked-data CRC
      * for later.
      */
-    if(rnc_crc(input, inputend-input) != bword(input - 4))
+    if(rnc_crc(input, static_cast<long>(inputend-input)) != bword(input - 4))
     {
 	    return RNC_PACKED_CRC_ERROR;
     }
